move serializers to own module
